Added pablo files with solution information
First attempt at making examples of the format for expressing information needed to numerically solve the model. I'm more confident in the discretization information and "problem format" than I am in the solution format. As presented, it states a method (EGM) and presents the necessary information. The alternative is to provide a set of directions or instructions in the solution block, rather than a method tag.