Add new modeling language draft document, example
I've overhauled and expanded the description of the proposed model specification language, incorporating (most of) our team discussions. The new version is specifically geared towards showing Pablo, almost being written as a letter to him. It includes a section walking through the most basic consumption-saving model specified in pablo (the language), which is included on this commit as BasicCSpablo.yaml.